# contracts/openapi/platform.yaml
openapi: 3.0.3
info:
  title: BuildingOS platform foundation API
  version: "0.1.0"
  description: >
    Exact operational routes for the BOS-001 platform foundation milestone.
    Service metadata plus BOS-010 F5a revenue APIs (subscription-service) and F2 building
    application & lifecycle APIs (building-service).
servers:
  - url: http://localhost:8080
    description: Local gateway
paths:
...
  /api/v1/building-applications/{applicationId}/documents:
    get:
      summary: "Verification document metadata"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    post:
      summary: "Upload a verification document (applicant, while editable)"
      description: >
        Stored in S3-compatible storage (D-28). Type detected from bytes; file name is display metadata only.
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/DocumentUpload" }
      responses:
        "201":
          description: "Created envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: "NOT_EDITABLE or DOCUMENT_LIMIT_REACHED"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "413":
          description: "DOCUMENT_TOO_LARGE"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "415":
          description: "UNSUPPORTED_DOCUMENT_TYPE (PDF, JPEG, PNG by content)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "503":
          description: "DEPENDENCY_UNAVAILABLE (object storage)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/building-applications/{applicationId}/documents/{documentId}:
    get:
      summary: "Download a document (applicant or platform admin)"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
        - name: documentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "Document bytes, Content-Disposition attachment"
          content:
            application/octet-stream:
              schema: { $ref: "#/components/schemas/DocumentBytes" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND or DOCUMENT_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "503":
          description: "DEPENDENCY_UNAVAILABLE"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    delete:
      summary: "Remove a document (applicant, while editable)"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
        - name: documentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: "Removed"
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND or DOCUMENT_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: "NOT_EDITABLE"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /api/v1/platform/building-applications/{applicationId}/notes:
    get:
      summary: "Internal notes (never shown to applicant)"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    post:
      summary: "Add an internal note"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteRequest" }
      responses:
        "201":
          description: "Created envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiEnvelope:
      type: object
      required: [success, data, traceId]
      properties:
        success: { type: boolean }
        data: {}
        meta: { type: object }
        traceId: { type: string }
...
    DocumentUpload:
      type: object
      required: [file]
      properties:
        file: { type: string, format: binary }
    DocumentBytes:
      type: string
      format: binary
...
    NoteRequest:
      type: object
      required: [body]
      properties:
        body: { type: string, maxLength: 2000 }
...
    ApiError:
      type: object
      required: [success, code, message, traceId]
      properties:
        success: { type: boolean, enum: [false] }
        code:
          type: string
          description: Stable error code; each operation lists the codes it may return.
          example: ACCESS_DENIED
        message: { type: string }
        traceId: { type: string, format: uuid }